requirements.txt
# requirements.txt

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0

# Database
supabase>=1.0.0

# Technical analysis
pandas-ta>=0.3.14b

# Machine Learning (opsiyonel, gelecek için)
scikit-learn>=1.3.0
optuna>=3.0.0

# Data visualization (opsiyonel)
matplotlib>=3.7.0
plotly>=5.15.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# API rate limiting
ratelimit>=2.2.1

# Async support (gelecek için)
asyncio
aiohttp>=3.8.0

# Logging
loguru>=0.7.0

# Performance
numba>=0.57.0

# Statistics
scipy>=1.11.0

# Time series analysis
statsmodels>=0.14.0